pyproject.toml
[project]
name = "dynamic-transformer"
version = "0.2.0"
requires-python = ">=3.8"
dependencies = [
    "torch>=2.0.0",
    "transformers>=4.30.0",
    "accelerate>=0.20.0",
    "datasets>=2.0.0",
    "hydra-core>=1.3.0",
    "omegaconf>=2.3.0",
    "wandb>=0.15.0",
    "tqdm>=4.60.0",
]

[project.optional-dependencies]
dev = [
    "black",
    "ruff",
    "pytest",
]

[tool.black]
line-length = 100

[tool.ruff]
line-length = 100
select = ["E", "F", "W", "I"]